# openapi_contract.yaml
openapi: 3.0.3
info:
  title: To-Do Backend API
  version: 1.0.0
  description: |
    Formal OpenAPI-style contract for the FastAPI to-do backend.
    This document reflects the currently implemented routes and payload shapes.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /departments:
    post:
      tags: [departments]
      summary: Create department (admin only)
      operationId: createDepartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepartmentCreate'
      responses:
        '200':
          description: Department created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepartmentResponse'
        '409':
          $ref: '#/components/responses/Error409'
    get:
      tags: [departments]
      summary: List departments
      operationId: listDepartments
      responses:
        '200':
          description: Departments list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DepartmentResponse'
  /users/{user_id}/departments:
    put:
      tags: [departments]
      summary: Assign user departments (admin only)
      operationId: assignUserDepartments
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDepartmentAssignRequest'
      responses:
        '200':
          description: Department assignment updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDepartmentAssignResponse'
...
  /activities:
    post:
      tags: [activities]
      summary: Create activity (admin only)
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityCreate'
      responses:
        '200':
          description: Activity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
  /activities/{activity_id}:
    put:
      tags: [activities]
      summary: Update activity (admin only)
      operationId: updateActivity
      parameters:
        - $ref: '#/components/parameters/ActivityIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityUpdate'
      responses:
        '200':
          description: Activity updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityResponse'
    delete:
      tags: [activities]
      summary: Delete activity (admin only)
      operationId: deleteActivity
      parameters:
        - $ref: '#/components/parameters/ActivityIdPath'
      responses:
        '200':
          description: Activity deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /tasks/{task_id}/activities:
    get:
      tags: [activities]
      summary: List task activities
      operationId: listTaskActivities
      parameters:
        - $ref: '#/components/parameters/TaskIdPath'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Search'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/ActivityStatus'
        - in: query
          name: sub_task_id
          schema:
            type: integer
      responses:
        '200':
          description: Activity list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      in: query
      name: page
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      in: query
      name: page_size
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
    Search:
      in: query
      name: search
      required: false
      schema:
        type: string
    UserIdPath:
      in: path
      name: user_id
      required: true
      schema:
        type: integer
    TaskIdPath:
      in: path
      name: task_id
      required: true
      schema:
        type: integer
...
    ActivityIdPath:
      in: path
      name: activity_id
      required: true
      schema:
        type: integer
...
  responses:
...
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      required: [message, code, dev_message, request_id, path]
      properties:
        message:
          type: string
        code:
          type: string
        dev_message:
          type: string
        request_id:
          type: string
        path:
          type: string
        details:
          type: array
          items: {}
...
    ActivityStatus:
      type: string
      enum: [complete, not complete]
    UserReference:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
...
    DepartmentCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
    DepartmentResponse:
      type: object
      required: [id, name]
      properties:
        id:
          type: integer
        name:
          type: string
    UserDepartmentAssignRequest:
      type: object
      required: [department_ids]
      properties:
        department_ids:
          type: array
          items:
            type: integer
    UserDepartmentAssignResponse:
      type: object
      required: [message, department_ids]
      properties:
        message:
          type: string
        department_ids:
          type: array
          items:
            type: integer
...
    ActivityCreate:
      type: object
      required: [title, description, date, sub_task_id]
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        sub_task_id:
          type: integer
    ActivityUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        status:
          $ref: '#/components/schemas/ActivityStatus'
        sub_task_id:
          type: integer
    ActivityResponse:
      type: object
      required: [id, title, description, date, status, sub_task_id, created_by]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        date:
          type: string
          format: date
        status:
          type: string
        sub_task_id:
          type: integer
        created_by:
          $ref: '#/components/schemas/UserReference'
    ActivityListResponse:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ActivityResponse'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...